Forked cfengine/core
libntech was forked from CFEngine community at this commit SHA:

https://github.com/cfengine/core/commit/3b6af9850007fe9339db7d7fcc69f490acea5867

To see git history, blame, etc. before this point, checkout the commit in
cfengine/core above. This was done because the cfengine repository already had
15 000 commits, takes quite some time to clone and most of the history is not
related to the contents of libntech (only a small part of the CFEngine codebase).
This will also give the repo a "fresh" start in git and github statistics.
Over time these will show contributors to the new libntech library, not to the
rest of the cfengine codebase.

Thank you to everyone who contributed to CFEngine in the past:

https://github.com/cfengine/core/blob/3b6af9850007fe9339db7d7fcc69f490acea5867/AUTHORS

Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>