.clang-format: Changed column limit to 79
Our 78 limit was causing the copyright blob to break in a weird
place:

```
  This file is part of CFEngine 3 - written and maintained by Northern.tech
  AS.
```

(The full sentence is 79 characters including 2 space indentation).

I looked around and found that PEP8 recommends 79 characters, so
might as well use this as a "standard". PEP8 explains:

> Limit all lines to a maximum of 79 characters.
> Limiting the required editor window width makes it possible
> to have several files open side-by-side, and works well when
> using code review tools that present the two versions in
> adjacent columns.

Seems like a good enough standard as any.

Changelog: None
Ticket: CFE-2989
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>