Numbers with multiple dots now correctly emit errors in JSON parsing
The JSON spec allows for maximum 1 dot (decimal point) in a number.
This is invalid:

```
[1.2.3]
```

Previously, it would be parsed as `1.2.3`, stored as a string,
but cause problems later, when copying or trying to convert to
float / double.

The JSON parser now emits the following error:

```
Unable to parse json data as number, two or more dots (decimal points)
```

and stops parsing.

Ticket: CFE-3138
Changelog: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>