Added stack_base.c to EXTRA_DIST
Should be part of dist tarball, but not compiled directly.
This commit makes it possible to run `make dist`, and then
`./configure && make` inside the expanded tarball, without
errors.

`stack_base.c` is not in `_SOURCES`, because it includes
static functions and definitions which are needed in both
`stack.c` and `threaded_stack.c`. Compiling it directly
would cause warnings about unused static functions.

This is also explained in `stack_base.c`:

```
/** @file stack_base.c
 * @brief Shared code between threaded_stack.c and stack.c
 *
 * This file should be included, not compiled separately.
 */
```

Ticket: None
Changelog: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>