Improve logging in file locking
Failing to acquire a file lock is generally not an error (the
lock may be held by a different process). But the information
about the failure and exact reason may be useful for debugging.

Failures in some other operations related to file locking are
errors and should be logged as such.