Added enum JsonType and GetJsonType()
When validating json, you usually want to know whether an
element is exactly the type it should be. First checking
ElementType and then PrimitiveType / ContainterType is tedious,
makes the code harder to read, and more prone to error.

Could be cleaned up in many places, but this is enough for now.
I want to use this in `classfilterjson()`.

Changelog: Title
Ticket: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>