Add a function for checking whether Log() would log a message
Log() does most of the heavy lifting of formatting the log
message like printf() only when the log level means the message
will actually be logged. But sometimes there is a lot of work
being done before Log() is even called. It is nice to have a
chance to tell if such work can be skipped because Log() would
just discard the result anyway.