Cleaned up IP address comparison functions
More cases of functions and variables which should be bool.
AFAICT the -1 return value was unreachable anyway. (And it was
treated as bool in the calling location).

Also refactored IPAddressCompareLess() to follow code style,
fixed indentation, added assertions and comments to make it
easier to understand.

Reported by LGTM:
https://lgtm.com/rules/1508604166004/

Changelog: None
Ticket: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>