Use a different name for the DATADIR macro
Unfortunately, on mingw builds, 'DATADIR' is a name of an enum
type. So we cannot use the same string as a name for our own
macro. Let's use CF_DATADIR as the macro name, but let's keep
other things similar to other directories we work with.

Ticket: ENT-6789
Changelog: None