Add a process manager "class"
In many places there's a need to keep track of sub-processes
spawned by the main process. Such processes need to be looked up
by various parameters and interacted with.

Having a common, reliable and fast implementation with high test
coverage that holds the various pieces of information about
sub-processes together is beneficial.

In the future, we should extend this new API with the process
spawning and termination functionality.

Ticket: CFE-3572
Changelog: None