Fix a security issue in the buffered logging implementation
If the message from the buffer is logged, it must not be
interpreted as a format string. Otherwise logging a message like
"%s GOT YOUR SECRETS,HAHA" into the buffer can then cause a more
or less random piece of memory to be logged when the buffer is
committed. This is the essence of the problem:

  char *str;
  asprintf(&str, "%s", "%s GOT YOUR SECRETS, HAHA");
  Log(LOG_LEVEL_ERR, str);

when Log() is called, it expects another argument providing the
replacement for the "%s" sequence at the beginning of 'str' and
so it finds some. Either causing a memory violation (good) or
logging some more or less random piece of memory (REALLY BAD).

Ticket: CFE-3708
Changelog: None