JSON parser now complains about trailing bogus data
The JSON parser used to stop the parsing of a string once a JSON element
was terminated. The remaining string was ignored. Now the parser checks
that the remaining string only contains whitespace characters, returning
a `JSON_PARSE_ERROR_INVALID_END` error code if this is not the case.

Ticket: CFE-4080
Changelog: Title
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>