Added CompareVersionExpression() for comparing 2 version numbers
Will be used in the upcoming CFEngine policy function; `version_compare()`.

This function takes 3 arguments, 2 version numbers and an
operator between them, thus resembling an expression to evaluate.
(Is this version bigger than this version, etc.)

Internally it uses the exact same logic as we use for comparing the
currently installed CFEngine version, for the cf_version_minimum
and similar functions and macros.

Ticket: CFE-3991
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>