Define type Regex and implement RegexDestroy()
When migrating from PCRE to PCRE2 we had to make a change
breaking API because all the functions working with REs were
using the PCRE's type `pcre`. To avoid this change in the future,
we define our own type `Regex` which we can later redefine
without causing any API breakage.

This also requires a definition of `RegexDestroy()` for use
instead of `pcre2_code_free()`.

Ticket: ENT-10629
Changelog: API working with regular expressions uses a custom
           type Regex instead of pcre2_code directly