Fixed bug where basename(3) modified its input string
It was assumed that the potential mutation of the path (the input
string) passed as an argument to basename(3) did not matter. However,
this assumption was wrong, as the path is later used to determine
whether or not the path points to a directory, using stat(2). In other
words the mutation of the path caused stat(2) to fail, due to the
basename being removed from it.

Ticket: ENT-11434
Changelog: None
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>