Fixed compiler warning -Werror=enum-int-mismatch
The function LogToSystemLog now takes an int as the log level, although
the desired type would be the enum "LogLevel". For some C compilers, the
last required argument must be self-promoting: that is, the default
promotions must not change its type (this is actually an ISO C
requirement). However, enums promote to an int. Thus, we cannot use them
and must use int instead.

Ticket: None
Changelog: None
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>
Co-authored-by: Craig Comstock <craig.comstock@northern.tech>