Fixed bug where `cf-net.exe get` fails to create file
When calling `fdopen()` in `safe_fopen_create_perms()` using the `"wx"`
mode we get...

```
warning: Invalid parameter passed to C runtime function.
```

...on Windows. Looking in both Linux man pages and Win32 API, I see that
the mode argument `"x"` is not specified as a valid argument in either
of them. Hence, it is probably not supported. Maybe Unix systems simply
ignore it, while Windows returns error? Either way, I believe it is safe
to omit it, given that we already called `open()` with the `O_EXCL`
flag, and that we are getting the file stream from that same file
descriptor.

Ticket: ENT-12511
Changelog: Title
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>