Introduce first bits of CMake support
CMake is a very common build system. While it can utilize
autotools-based projects as sub-components, it's much better to
have a CMake-native support and avoid the requirement of having
working autotools available.

This is just the very first bit kickstarting the effort and
setting up the base for a large future expansion.

Ticket: MEN-7819
Changelog: libntech now has a limitted support of CMake, expected
           to become more complete in the future