Sparse file copying no longer aborts if fallocate is unsupported
On my Ubuntu 24 VM, I experienced that FileSparseCopy() kept getting
aborted with the following error message:

```
   error: Failed to pre-allocate space for 'promises.cf.cfnew': Operation not supported
```

In the man pages I saw that the system call fallocate(2) can fail if the
file system does not support it. However, fallocate(2) is nothing more
than an optimization. It is not strictly needed. Hence, the file copying
should proceed if fallocate(2) is unsupported.

Changelog: Title
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>